Refactor some of the CDC/changelog magic out.
